requirements.txt
appium-python-client==3.0.0
selenium==4.21.0
pytest
